fix: 아이디 , 닉네임 유효성 검증 로직 수정
아이디 , 닉네임 유효성 검증 로직 수정
