+++ src/main/java/com/team/jjan/join/dto/ValidationRequest.java
@@ -1,0 +1,12 @@
+package com.team.jjan.join.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ValidationRequest {
+    String data;
+}
